# Conflicts:
#	.idea/gradle.xml

+++ backend/src/main/kotlin/app/ehrenamtskarte/backend/application/webservice/EakApplicationMutationService.kt
@@ -1,10 +1,15 @@
 package app.ehrenamtskarte.backend.application.webservice
 
+import app.ehrenamtskarte.backend.application.database.EakApplicationEntity
 import app.ehrenamtskarte.backend.application.database.repos.EakApplicationRepository
 import app.ehrenamtskarte.backend.application.webservice.schema.create.BlueCardApplication
-import app.ehrenamtskarte.backend.application.webservice.schema.create.GoldenEakCardApplication
+import app.ehrenamtskarte.backend.application.webservice.schema.create.GoldenCardApplication
+import app.ehrenamtskarte.backend.auth.database.AdministratorEntity
+import app.ehrenamtskarte.backend.auth.service.Authorizer.mayDeleteApplicationsInRegion
 import app.ehrenamtskarte.backend.common.webservice.GraphQLContext
-import com.expediagroup.graphql.annotations.GraphQLDescription
+import app.ehrenamtskarte.backend.common.webservice.UnauthorizedException
+import com.expediagroup.graphql.generator.annotations.GraphQLDescription
+import graphql.schema.DataFetchingEnvironment
 
 @Suppress("unused")
 class EakApplicationMutationService {
@@ -13,28 +18,51 @@ class EakApplicationMutationService {
     fun addBlueEakApplication(
         regionId: Int,
         application: BlueCardApplication,
-        graphQLContext: GraphQLContext
+        dfe: DataFetchingEnvironment
     ): Boolean {
-        EakApplicationRepository.addEakApplication(regionId, application, graphQLContext, EakApplicationRepository::validateBlueApplication)
+        val context = dfe.getLocalContext<GraphQLContext>()
+        EakApplicationRepository.addEakApplication(
+            regionId,
+            application,
+            context,
+            EakApplicationRepository::validateBlueApplication
+        )
         return true
     }
 
     @GraphQLDescription("Stores a new golden digital EAK")
     fun addGoldenEakApplication(
         regionId: Int,
-        application: GoldenEakCardApplication,
-        graphQLContext: GraphQLContext
+        application: GoldenCardApplication,
+        dfe: DataFetchingEnvironment
     ): Boolean {
-        EakApplicationRepository.addEakApplication(regionId, application, graphQLContext, EakApplicationRepository::validateGoldenApplication)
+        val context = dfe.getLocalContext<GraphQLContext>()
+        EakApplicationRepository.addEakApplication(
+            regionId,
+            application,
+            context,
+            EakApplicationRepository::validateGoldenApplication
+        )
         return true
     }
 
     @GraphQLDescription("Deletes the application with specified id")
     fun deleteApplication(
-        context: GraphQLContext,
-        applicationId: Int
+        applicationId: Int,
+        dfe: DataFetchingEnvironment
     ): Boolean {
-        context.enforceSignedIn()
+        val context = dfe.getLocalContext<GraphQLContext>()
+        val jwtPayload = context.enforceSignedIn()
+
+        val application = EakApplicationEntity.findById(applicationId) ?: throw UnauthorizedException()
+        // We throw an UnauthorizedException here, as we do not know whether there was an application with id
+        // `applicationId` and whether this application was contained in the user's project & region.
+
+        val user = AdministratorEntity.findById(jwtPayload.userId)
+        if (!mayDeleteApplicationsInRegion(user, application.regionId.value)) {
+            throw UnauthorizedException()
+        }
+
         return EakApplicationRepository.delete(applicationId, context)
     }
 }
